Implement Contraint, Restrict on Delete of any parent table in demographic information

+++ app/Http/Controllers/Api/V1/Admin/LocationController.php
@@ -350,9 +350,18 @@ class LocationController extends Controller
         $validator->validated();
 
         $division = Location::whereId($id)->first();
+
+        // check if division has any child if yes then return exception else delete
+        if($division->children->count() > 0){
+
+            return $this->sendError('you cannot delete this because it has child', [], 500);
+        }
+
+
         if($division){
             $division->delete();
         }
+
         activity("Division")
         ->causedBy(auth()->user())
         ->log('Division Deleted!!');
@@ -737,6 +746,10 @@ class LocationController extends Controller
         $validator->validated();
 
         $district = Location::whereId($id)->whereType($this->district)->first();
+        if($district->children->count() > 0){
+
+            return $this->sendError('you cannot delete this because it has child', [], 500);
+        }
         if($district){
             $district->delete();
         }
@@ -1146,6 +1159,10 @@ class LocationController extends Controller
         $validator->validated();
 
         $city = Location::whereId($id)->whereType($this->city)->first();
+        if($city->children->count() > 0){
+
+            return $this->sendError('you cannot delete this because it has child', [], 500);
+        }
         if($city){
             $city->delete();
         }
@@ -1606,6 +1623,10 @@ class LocationController extends Controller
         $validator->validated();
 
         $thana = Location::whereId($id)->whereType($this->thana)->first();
+        if($thana->children->count() > 0){
+
+            return $this->sendError('you cannot delete this because it has child', [], 500);
+        }
         if($thana){
             $thana->delete();
         }
@@ -2009,6 +2030,10 @@ class LocationController extends Controller
         $validator->validated();
 
         $union = Location::whereId($id)->whereType($this->union)->first();
+        if($union->children->count() > 0){
+
+            return $this->sendError('you cannot delete this because it has child', [], 500);
+        }
         if($union){
             $union->delete();
         }
@@ -2443,6 +2468,10 @@ class LocationController extends Controller
         $validator->validated();
 
         $ward = Location::whereId($id)->whereType($this->ward)->first();
+        if($ward->children->count() > 0){
+
+            return $this->sendError('you cannot delete this because it has child', [], 500);
+        }
         if($ward){
             $ward->delete();
         }
